Fix a bug in CollectionIterator.
This library was born from a collection of usefull and common code used during my PhD.
During this time is grown quite a bit, also becoming quite complex.
Documentation and unit tests will be added to the code as I continue to work in this lib.

Several bits and pieces of this lib can be usefull for some other students.