Refactorign SearchEngine code.
Search engine was promoted to a Interface.
Most of the code was transfered to WebSearchEngine.
Abstract class that helps with SearchEngines based on web APIs.
Each WebSearchEngine (Faroo, Yacy and Searx) needs only to implement a simple iterator to parse a single http request.
Most of the work will be process a single JsonArray.
MaxResults is therefore implemented in all WebSearchEngines, improving stoping criteria.
Moreover, the code responsible to retrieve the ResultsIterator was moved to hasNext() functions instead of the construction.
Decreasing creation time and improving performance (HTTP request can be slow).