Fix minor issue with unit test.
Start the work to reduce overhead with matrix multiplication.
It will increase speed for A*Bt (second matrix is transposed).
The implemented algorithm uses a fast transpose to create a aligned B matrix. If B is to be transpoted only for the multiplication we can save that operation and achieve good performance.