Univariate regression was moved to ml library.
ArrayUtils library is becoming quite big.
The primary consequence is that is becomes difficult to manage it.
In order to solve this issue, the code will be broken into smaller and more focused libraries.

The first step was to move the regression code to ml library.
As an improvement, the methods now return an Class that contains data regarding the regression (instead of a triplet of doubles).

The classes have implemented a solve method that computes values with the model. And a correct implementation of: toString, equals and hasdCode.