This commit closes issue #2.
Whenever one of the means is zero (having no elements), the threshold
is defined as the previous one to finnish the cycle.

Since this should only happen when the elements are all the same.

Other possibility would be to use a ternary operator and compute the new
threshold in the following manner:

    mat = (cat > 0) ? mat / cat: 0.0;
    mbt = (cbt > 0) ? mbt / cbt: 0.0;
    t = (mat + mbt) / 2.0;

However this may lead to a threshold that is outside of the range from
the original data.
As such, the first solution was select as the fix.