Fixes issue #5.
Several improvements.
Added an iterative version of qselect (instead of recursive).
This is used to find the median of an array in linear time (O(n)).
The median code has some optimizations:
1. Based on qselect.
2. Special cases for arrays with size 1,2 and 3.
3. Copies elements to a buffer in order to not modify the original array.

The median is begin used by DBSCAN method from ML library.
If there is no evident elbow, the median of the distances can be used.