Added a Dynamic Thread Pool.
A new worker (IdleWorker) was developed. It stores the time it last worked. With that is possible to count the idle time.
If a worker is idle for more than a predefined number of seconds, it is stopped and deallocate.
Whenever necessary the worker will be awaken to resume working.